got rid of dist, moved to ./

+++ uploader.js
@@ -8,7 +8,9 @@ exports.uploader = undefined;
 var _extends = Object.assign || function (target) { for (var i = 1; i < arguments.length; i++) { var source = arguments[i]; for (var key in source) { if (Object.prototype.hasOwnProperty.call(source, key)) { target[key] = source[key]; } } } return target; };
 
 var uploader = exports.uploader = function () {
-	var _ref = _asyncToGenerator( /*#__PURE__*/regeneratorRuntime.mark(function _callee(sourcePath, targetPath, r_oss, bucket, deleteBool, options) {
+	var _ref = _asyncToGenerator( /*#__PURE__*/regeneratorRuntime.mark(function _callee(sourcePath, targetPath, r_oss, bucket) {
+		var deleteBool = arguments.length > 4 && arguments[4] !== undefined ? arguments[4] : false;
+		var options = arguments[5];
 		var oss, ossStream;
 		return regeneratorRuntime.wrap(function _callee$(_context) {
 			while (1) {
@@ -34,8 +36,10 @@ var uploader = exports.uploader = function () {
 							});
 							upload.on('uploaded', function () {
 								console.log('Upload complete.');
-								_fs2.default.unlinkSync(sourcePath);
-								console.log('Deleted original file (' + sourcePath + ')');
+								if (deleteBool) {
+									_fs2.default.unlinkSync(sourcePath);
+									console.log('Deleted original file (' + sourcePath + ')');
+								}
 								_fs2.default.unlinkSync(newSourcePath);
 								console.log('Deleted generated compressed file (' + newSourcePath + ')');
 								resolve();
@@ -52,7 +56,7 @@ var uploader = exports.uploader = function () {
 		}, _callee, this);
 	}));
 
-	return function uploader(_x, _x2, _x3, _x4, _x5, _x6) {
+	return function uploader(_x, _x2, _x3, _x4) {
 		return _ref.apply(this, arguments);
 	};
 }();
